requirements.txt
gspread
environs
psycopg2
requests
django-cors-headers